copy old demo files (Demo.xcodeproj) into new demo target (Moya.xcodeproj)
