--- Demo/Demo/AppDelegate.swift
@@ -1,9 +1,0 @@
-import UIKit
-
-@UIApplicationMain
-class AppDelegate: UIResponder, UIApplicationDelegate {
-
-    var window: UIWindow?
-
-}
-
