--- Demo/Demo/Base.lproj/Main.storyboard
@@ -1,87 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="10116" systemVersion="15F34" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" initialViewController="9fE-rm-iIL">
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="10085"/>
-        <capability name="Navigation items with more than one left or right bar item" minToolsVersion="7.0"/>
-    </dependencies>
-    <scenes>
-        <!--Root View Controller-->
-        <scene sceneID="aLk-ch-wh3">
-            <objects>
-                <tableViewController id="iag-78-iuI" customClass="ViewController" customModule="Demo" customModuleProvider="target" sceneMemberID="viewController">
-                    <tableView key="view" clipsSubviews="YES" contentMode="scaleToFill" alwaysBounceVertical="YES" dataMode="prototypes" style="plain" separatorStyle="default" rowHeight="44" sectionHeaderHeight="22" sectionFooterHeight="22" id="XqI-te-CgY">
-                        <rect key="frame" x="0.0" y="0.0" width="600" height="600"/>
-                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-                        <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
-                        <prototypes>
-                            <tableViewCell contentMode="scaleToFill" selectionStyle="default" indentationWidth="10" reuseIdentifier="Cell" textLabel="a51-T5-z9Q" style="IBUITableViewCellStyleDefault" id="vDB-kj-Hda">
-                                <rect key="frame" x="0.0" y="86" width="600" height="44"/>
-                                <autoresizingMask key="autoresizingMask"/>
-                                <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" tableViewCell="vDB-kj-Hda" id="Azh-KN-cNV">
-                                    <rect key="frame" x="0.0" y="0.0" width="600" height="43.5"/>
-                                    <autoresizingMask key="autoresizingMask"/>
-                                    <subviews>
-                                        <label opaque="NO" multipleTouchEnabled="YES" contentMode="left" text="Title" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" id="a51-T5-z9Q">
-                                            <rect key="frame" x="15" y="0.0" width="570" height="43.5"/>
-                                            <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
-                                            <fontDescription key="fontDescription" type="system" pointSize="16"/>
-                                            <color key="textColor" red="0.0" green="0.0" blue="0.0" alpha="1" colorSpace="calibratedRGB"/>
-                                            <nil key="highlightedColor"/>
-                                        </label>
-                                    </subviews>
-                                </tableViewCellContentView>
-                            </tableViewCell>
-                        </prototypes>
-                        <connections>
-                            <outlet property="dataSource" destination="iag-78-iuI" id="g7D-Xm-aDY"/>
-                            <outlet property="delegate" destination="iag-78-iuI" id="VUp-HU-CZa"/>
-                        </connections>
-                    </tableView>
-                    <navigationItem key="navigationItem" title="Root View Controller" id="fZ1-y1-Xxt">
-                        <leftBarButtonItems>
-                            <barButtonItem title="Zen" id="bkk-cH-CrR">
-                                <connections>
-                                    <action selector="zenWasPressed:" destination="iag-78-iuI" id="cAQ-yX-VOU"/>
-                                </connections>
-                            </barButtonItem>
-                            <barButtonItem title="Giphy" id="O45-Iq-njJ">
-                                <connections>
-                                    <action selector="giphyWasPressed:" destination="iag-78-iuI" id="fNG-cI-scK"/>
-                                </connections>
-                            </barButtonItem>
-                            <barButtonItem title="Download" id="ofd-BP-GkV">
-                                <connections>
-                                    <action selector="downloadWasPressed:" destination="iag-78-iuI" id="xo5-T3-yJG"/>
-                                </connections>
-                            </barButtonItem>
-                        </leftBarButtonItems>
-                        <barButtonItem key="rightBarButtonItem" systemItem="search" id="ezq-hH-fx4">
-                            <connections>
-                                <action selector="searchWasPressed:" destination="iag-78-iuI" id="BQa-s5-8DX"/>
-                            </connections>
-                        </barButtonItem>
-                    </navigationItem>
-                </tableViewController>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="n3h-Zg-pHf" userLabel="First Responder" sceneMemberID="firstResponder"/>
-            </objects>
-            <point key="canvasLocation" x="753" y="338"/>
-        </scene>
-        <!--Navigation Controller-->
-        <scene sceneID="OcX-U2-LWP">
-            <objects>
-                <navigationController id="9fE-rm-iIL" sceneMemberID="viewController">
-                    <navigationBar key="navigationBar" contentMode="scaleToFill" id="569-cx-LCm">
-                        <rect key="frame" x="0.0" y="0.0" width="320" height="44"/>
-                        <autoresizingMask key="autoresizingMask"/>
-                    </navigationBar>
-                    <connections>
-                        <segue destination="iag-78-iuI" kind="relationship" relationship="rootViewController" id="7qq-fc-Be8"/>
-                    </connections>
-                </navigationController>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="uGI-Ha-0EF" userLabel="First Responder" sceneMemberID="firstResponder"/>
-            </objects>
-            <point key="canvasLocation" x="-67" y="338"/>
-        </scene>
-    </scenes>
-</document>
